+++ HISTORY.md
@@ -3,5 +3,12 @@ History
 
 - 0.9 (31-01-2014)
 	- Initial public release
+
 - 1.0-RC (06-02-2014)
-	- Initial public release
+	- Upload from remote URL
+	- Upload GitHub packages
+	- Bugfixes
+
+- 1.0-RC2 (11-02-2014)
+	- Delete installed extras directly
+	- Bugfixes